clan/cli: Use correct help formatter
Use the correct help formatter, so that our spacing is preserved.
This brings it in line with the rest of the subcommands.

Instead of:
```
This command prints the metadata of a clan. Examples: $ clan show
--flake [PATH] Name: My Empty Clan Description: some
nice description Icon: A path to the png
```
In one single line, we now get:
```
Examples:

  $ clan show --flake [PATH]
  Name: My Empty Clan
  Description: some nice description
  Icon: A path to the png
```
Which is easier to read and parse, because it has structure to it.