pkgs/cli/machines: Add `metavar` check
Add a `metavar` check for the machines subcommand to accomodate the
following https://github.com/python/cpython/issues/67037 workaround.

This test ensures that new subcommands will be documented by the CLI and
the documentation site and will error out if that is not the case.

If subcommands are added that should be excluded on purpose, they should
be added to the `ignored_subcommands` list.